refactor(api): move /api/chapter/login to /api/chapters/login
